requirements.txt
PyGithub>=1.55,<2.0
pyyaml>=6.0
python-dotenv>=0.19.2
schedule>=1.1.0
openai>=0.27.0
Flask>=2.0.1
click>=8.0.3
pytz>=2021.3
matplotlib>=3.4.3
flask-talisman>=0.8.1